The diff shows changes in 3 files: TambahDestinasi.php, tabel_paket.php, and TambahDestinasi.css.
In TambahDestinasi.php, the path to the logo image was changed from "/assets/LogoM360.png" to "../assets/LogoM360.png". Additionally, the input type for the "Batal" button was changed from "reset" to "submit".

In tabel_paket.php, the data type of the "gambar" column was changed from VARCHAR(255) to BLOB.

In style/TambahDestinasi.css, the "deskripsi" class was added to the list of classes that have a margin-bottom of 5%.